+++ src/main/java/http/method/GetHandler.java
@@ -1,0 +1,4 @@
+package http.method;
+
+public class GetHandler {
+}
