+++ src/main/java/http/method/PostHandler.java
@@ -1,0 +1,4 @@
+package http.method;
+
+public class PostHandler {
+}
